+++ client/functions/enlaceTx.py
