+++ client/temp/aplicacao.py
